Update README.md
For each entity, they relate back to the topic of running out of gas. With the first solution being call 911, then walk to the nearest gas station, next wait to see who will help, and the lastly call roadside assistance being the solution. Each solution was a process of elimination on which would be the best solution. Being that if traveling to CA and some parts may not be walking distance from things, and other parts not being safe; roadside assistance was the best solution.